3765: Conditional Logic: Move ListStatus enum to new file
