+++ grc/birth_registration/__init__.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from flask import Blueprint, render_template, request, url_for, session
-from grc.models import ListStatus
+from grc.list_status import ListStatus
 from grc.birth_registration.forms import  NameForm, DobForm, UkCheckForm, CountryForm, PlaceOfBirthForm, MothersNameForm, FatherNameCheckForm, FathersNameForm, AdoptedForm, AdoptedUKForm, ForcesForm, CheckYourAnswers
 from grc.utils.decorators import LoginRequired
 from grc.utils.application_progress import save_progress
